Update Basic example to latest dependencies
Update dependency versions and project files for iOS and Android
