+++ examples/Basic/android/app/build.gradle
@@ -55,7 +55,13 @@ import com.android.build.OutputFile
  *   // date; if you have any other folders that you want to ignore for performance reasons (gradle
  *   // indexes the entire tree), add them here. Alternatively, if you have JS files in android/
  *   // for example, you might want to remove it from here.
- *   inputExcludes: ["android/**", "ios/**"]
+ *   inputExcludes: ["android/**", "ios/**"],
+ *
+ *   // override which node gets called and with what additional arguments
+ *   nodeExecutableAndArgs: ["node"]
+ *
+ *   // supply additional arguments to the packager
+ *   extraPackagerArgs: []
  * ]
  */
 
@@ -89,7 +95,7 @@ android {
         ndk {
             abiFilters "armeabi-v7a", "x86"
         }
-        renderscriptTargetApi 20
+        renderscriptTargetApi 21
         renderscriptSupportModeEnabled true
     }
     splits {
@@ -126,10 +132,10 @@ repositories {
 }
 
 dependencies {
+    compile project(':react-native-blur')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.4.0"
-    compile "com.facebook.react:react-native:+"  // From node_modules
-    compile project(':react-native-blur')
+    compile "com.facebook.react:react-native:0.36.0"  // From node_modules
 }
 
 buildscript {
