+++ examples/Basic/android/app/src/main/AndroidManifest.xml
@@ -4,12 +4,13 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+      android:name=".MainApplication"
       android:allowBackup="true"
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
       android:theme="@style/AppTheme">
       <activity
-        android:name=".MainActivity"
+        android:name="com.basic.MainActivity"
         android:label="@string/app_name">
         <intent-filter>
             <action android:name="android.intent.action.MAIN" />
